fix event alignment for 32bit platforms

+++ lib/src/async/async_io.rs
@@ -181,14 +181,14 @@ impl AsyncRequest {
     ///    .as_input()
     ///    .with_edge_detection(gpiocdev::line::EdgeDetection::BothEdges)
     ///    .request()?;
-    /// let mut buf = vec![0; req.edge_event_size() * 3];
+    /// let mut buf = vec![0_u64; req.edge_event_u64_size() * 3];
     /// let areq = AsyncRequest::new(req);
     /// let num_evts = areq.read_edge_events_into_slice(&mut buf).await?;
     /// // process events in buf...
     /// # Ok(())
     /// # }
     /// ```
-    pub async fn read_edge_events_into_slice(&self, buf: &mut [u8]) -> Result<usize> {
+    pub async fn read_edge_events_into_slice(&self, buf: &mut [u64]) -> Result<usize> {
         loop {
             self.0.readable().await?;
             let req = self.0.get_ref();
